# clients/rust/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@flowcatalyst.tech
    name: FlowCatalyst Support
    url: https://flowcatalyst.tech/support
  description: Control plane API for FlowCatalyst - multi-tenant event routing platform.
  license:
    name: Proprietary
    url: https://flowcatalyst.tech/license
  title: FlowCatalyst Platform API
  version: "1.0"
paths:
...
  /admin/platform/auth-configs:
    get:
      description: Returns all authentication configurations
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/internal_platform_api.AuthConfigDTO"
                type: array
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: List auth configs
      tags:
        - Auth Config
    post:
      description: Creates a new authentication configuration for an email domain
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_platform_api.CreateAuthConfigRequest"
        description: Auth config to create
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.AuthConfigDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Create auth config
      tags:
        - Auth Config
  "/admin/platform/auth-configs/{id}":
    delete:
      description: Deletes an authentication configuration
      parameters:
        - description: Auth config ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Delete auth config
      tags:
        - Auth Config
    get:
      description: Returns a specific authentication configuration
      parameters:
        - description: Auth config ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.AuthConfigDTO"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Get auth config by ID
      tags:
        - Auth Config
    put:
      description: Updates an existing authentication configuration
      parameters:
        - description: Auth config ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_platform_api.CreateAuthConfigRequest"
        description: Auth config updates
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.AuthConfigDTO"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_platform_api.ErrorResponse"
      security:
        - BearerAuth: []
      summary: Update auth config
      tags:
        - Auth Config
...
servers:
  - url: //localhost:8080/api
components:
  securitySchemes:
    BearerAuth:
      description: 'JWT Bearer token. Format: "Bearer {token}"'
      in: header
      name: Authorization
      type: apiKey
    SessionCookie:
      description: Session cookie for browser-based authentication
      in: cookie
      name: FLOWCATALYST_SESSION
      type: apiKey
  schemas:
...
    internal_platform_api.AuthConfigDTO:
      properties:
        additionalClientIds:
          items:
            type: string
          type: array
        authProvider:
          type: string
        configType:
          type: string
        createdAt:
          type: string
        emailDomain:
          type: string
        entraTenantId:
          type: string
        grantedClientIds:
          items:
            type: string
          type: array
        groupsClaim:
          type: string
        id:
          type: string
        idpType:
          type: string
        oidcClientId:
          type: string
        oidcIssuerPattern:
          type: string
        oidcIssuerUrl:
          type: string
        oidcMultiTenant:
          type: boolean
        primaryClientId:
          type: string
        rolesClaim:
          type: string
        updatedAt:
          type: string
      type: object
...
    internal_platform_api.CreateAuthConfigRequest:
      properties:
        additionalClientIds:
          items:
            type: string
          type: array
        authProvider:
          type: string
        configType:
          type: string
        emailDomain:
          type: string
        entraTenantId:
          type: string
        grantedClientIds:
          items:
            type: string
          type: array
        groupsClaim:
          type: string
        idpType:
          type: string
        oidcClientId:
          type: string
        oidcClientSecretRef:
          type: string
        oidcIssuerPattern:
          type: string
        oidcIssuerUrl:
          type: string
        oidcMultiTenant:
          type: boolean
        primaryClientId:
          type: string
        rolesClaim:
          type: string
      type: object
...
    internal_platform_api.ErrorResponse:
      properties:
        details:
          type: string
        error:
          type: string
        message:
          type: string
      type: object
...